scripts/requirements.txt
PyYAML
numpy
pandas
scipy
requests
